Press 'r' to restart (and new college sprites)

+++ core/src/yorkpirates/objects/Ship.java
@@ -93,9 +93,9 @@ public abstract class Ship implements GameObject, HasTransform {
         // Prevent the ship from overlapping colleges
         for (int i = 0; i<game.colleges.length; i++) {
             Rectangle currentCollege = game.colleges[i].transform;
-            if (targetPosition.x >= currentCollege.x-(float)currentCollege.width/2 &&
+            if (targetPosition.x >= currentCollege.x- currentCollege.width /2 &&
                     targetPosition.x <= currentCollege.x+currentCollege.width &&
-                    targetPosition.y >= currentCollege.y-(float)currentCollege.height/2 &&
+                    targetPosition.y >= currentCollege.y- currentCollege.height /2 &&
                     targetPosition.y <= currentCollege.y+currentCollege.height) {
                 targetPosition = transform.getPosition(new Vector2());
             }
